use new hyperlane cosmos with couple changes (moved stuff to test files)
